+++ WebShop.Clients/Services/Users/UserTwoFactorStoreClient.cs
@@ -1,0 +1,28 @@
+﻿using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using WebShop.Domain.Entities;
+
+namespace WebShop.Clients.Services.Users
+{
+    public partial class UsersClient
+    {
+        #region IUserTwoFactorStore
+
+        public async Task SetTwoFactorEnabledAsync(ApplicationUser user, bool enabled, CancellationToken cancellationToken)
+        {
+            user.TwoFactorEnabled = enabled;
+            var url = $"{ServiceAddress}/setTwoFactorEnabled/{enabled}";
+            await PostAsync(url, user);
+        }
+
+        public async Task<bool> GetTwoFactorEnabledAsync(ApplicationUser user, CancellationToken cancellationToken)
+        {
+            var url = $"{ServiceAddress}/getTwoFactorEnabled";
+            var result = await PostAsync(url, user);
+            return await result.Content.ReadAsAsync<bool>();
+        }
+
+        #endregion
+    }
+}
