+++ WebShop.Clients/Services/Users/UsersClient.cs
@@ -1,19 +1,15 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Generic;
 using System.Net.Http;
-using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 using WebShop.Clients.Base;
-using WebShop.Domain.Dto.User;
 using WebShop.Domain.Entities;
 using WebShop.Interfaces.Api;
 
 namespace WebShop.Clients.Services.Users
 {
-    public class UsersClient : BaseClient, IUserIdentity
+    public partial class UsersClient : BaseClient, IUserIdentity
     {
         protected sealed override string ServiceAddress { get; set; }
 
@@ -21,252 +17,6 @@ namespace WebShop.Clients.Services.Users
         {
             ServiceAddress = "api/users";
         }
-
-        #region IUserRoleStore
-
-        public async Task AddToRoleAsync(ApplicationUser user, string roleName, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<IList<string>> GetRolesAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<IList<ApplicationUser>> GetUsersInRoleAsync(string roleName, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<bool> IsInRoleAsync(ApplicationUser user, string roleName, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task RemoveFromRoleAsync(ApplicationUser user, string roleName, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        #endregion
-
-        #region IUserClaimStore
-
-        public async Task AddClaimsAsync(ApplicationUser user, IEnumerable<Claim> claims, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<IList<Claim>> GetClaimsAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<IList<ApplicationUser>> GetUsersForClaimAsync(Claim claim, CancellationToken cancellationToken)
-        {
-           
-        }
-
-        public async Task RemoveClaimsAsync(ApplicationUser user, IEnumerable<Claim> claims, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task ReplaceClaimAsync(ApplicationUser user, Claim claim, Claim newClaim, CancellationToken cancellationToken)
-        {
-           
-        }
-
-        #endregion
-
-        #region IUserPasswordStore
-
-        public async Task SetPasswordHashAsync(ApplicationUser user, string passwordHash, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<string> GetPasswordHashAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-           
-        }
-
-        public async Task<bool> HasPasswordAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-           
-        }
-
-        #endregion
-
-        #region IUserTwoFactorStore
-
-        public async Task SetTwoFactorEnabledAsync(ApplicationUser user, bool enabled, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<bool> GetTwoFactorEnabledAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        #endregion
-
-        #region IUserEmailStore
-
-        public async Task SetEmailAsync(ApplicationUser user, string email, CancellationToken cancellationToken)
-        {
-            user.Email = email;
-            var url = $"{ServiceAddress}/setEmail/{email}";
-            await PostAsync(url, user);
-        }
-
-        public async Task<string> GetEmailAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-           
-        }
-
-        public async Task<bool> GetEmailConfirmedAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task SetEmailConfirmedAsync(ApplicationUser user, bool confirmed, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<ApplicationUser> FindByEmailAsync(string normalizedEmail, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task<string> GetNormalizedEmailAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        public async Task SetNormalizedEmailAsync(ApplicationUser user, string normalizedEmail, CancellationToken cancellationToken)
-        {
-            
-        }
-
-        #endregion
-
-        #region IUserPhoneNumberStore
-
-
-        public async Task SetPhoneNumberAsync(ApplicationUser user, string phoneNumber, CancellationToken cancellationToken)
-        {
-            user.PhoneNumber = phoneNumber;
-            var url = $"{ServiceAddress}/setPhoneNumber/{phoneNumber}";
-            await PostAsync(url, user);
-        }
-
-        public async Task<string> GetPhoneNumberAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/getPhoneNumber";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsStringAsync();
-        }
-
-        public async Task<bool> GetPhoneNumberConfirmedAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/getPhoneNumberConfirmed";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsAsync<bool>();
-        }
-
-        public async Task SetPhoneNumberConfirmedAsync(ApplicationUser user, bool confirmed, CancellationToken cancellationToken)
-        {
-            user.PhoneNumberConfirmed = confirmed;
-            var url = $"{ServiceAddress}/setPhoneNumberConfirmed/{confirmed}";
-            await PostAsync(url, user);
-        }
-
-        #endregion
-
-        #region IUserLockoutStore
-
-        public async Task<DateTimeOffset?> GetLockoutEndDateAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/getLockoutEndDate";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsAsync<DateTimeOffset?>();
-        }
-
-        public async Task SetLockoutEndDateAsync(ApplicationUser user, DateTimeOffset? lockoutEnd, CancellationToken cancellationToken)
-        {
-            user.LockoutEnd = lockoutEnd;
-            var url = $"{ServiceAddress}/setLockoutEndDate";
-            await PostAsync(url, new SetLockoutDto() { User = user, LockoutEnd = lockoutEnd });
-        }
-
-        public async Task<int> IncrementAccessFailedCountAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/incrementAccessFailedCount";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsAsync<int>();
-        }
-
-        public async Task ResetAccessFailedCountAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/resetAccessFailedCount";
-            await PostAsync(url, user);
-        }
-
-        public async Task<int> GetAccessFailedCountAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/getAccessFailedCount";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsAsync<int>();
-        }
-
-        public async Task<bool> GetLockoutEnabledAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/getLockoutEnabled";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsAsync<bool>();
-        }
-
-        public async Task SetLockoutEnabledAsync(ApplicationUser user, bool enabled, CancellationToken cancellationToken)
-        {
-            user.LockoutEnabled = enabled;
-            var url = $"{ServiceAddress}/setLockoutEnabled/{enabled}";
-            await PostAsync(url, user);
-        }
-
-        #endregion
-
-        #region IUserLoginStore
-
-        public async Task AddLoginAsync(ApplicationUser user, UserLoginInfo login, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/addLogin";
-            await PostAsync(url, new AddLoginDto() { User = user, UserLoginInfo = login });
-        }
-
-        public async Task RemoveLoginAsync(ApplicationUser user, string loginProvider, string providerKey, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/removeLogin/{loginProvider}/{providerKey}";
-            await PostAsync(url, user);
-        }
-
-        public async Task<IList<UserLoginInfo>> GetLoginsAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/getLogins";
-            var result = await PostAsync(url, user);
-            return await result.Content.ReadAsAsync<List<UserLoginInfo>>();
-        }
-
-        public async Task<ApplicationUser> FindByLoginAsync(string loginProvider, string providerKey, CancellationToken cancellationToken)
-        {
-            var url = $"{ServiceAddress}/user/findByLogin/{loginProvider}/{providerKey}";
-            return await GetAsync<ApplicationUser>(url);
-        }
-
-        #endregion
 
         #region IUserStore
 
@@ -342,11 +92,9 @@ namespace WebShop.Clients.Services.Users
 
         public void Dispose()
         {
-            throw new NotImplementedException();
+            client.Dispose();
         }
 
         #endregion
-
-
     }
 }
