+++ WebShop.ServicesHosting/Controllers/UserStore/UserPhoneNumberStore.cs
@@ -1,0 +1,37 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
+using WebShop.Domain.Entities;
+
+namespace WebShop.ServicesHosting.Controllers
+{
+    public partial class UsersApiController
+    {
+        #region IUserPhoneNumberStore
+
+        [HttpPost("getPhoneNumber")]
+        public async Task<string> GetPhoneNumberAsync([FromBody]ApplicationUser user)
+        {
+            return await _userStore.GetPhoneNumberAsync(user);
+        }
+
+        [HttpPost("getPhoneNumberConfirmed")]
+        public async Task<bool> GetPhoneNumberConfirmedAsync([FromBody]ApplicationUser user)
+        {
+            return await _userStore.GetPhoneNumberConfirmedAsync(user);
+        }
+
+        [HttpPost("setPhoneNumber/{phoneNumber}")]
+        public async Task SetPhoneNumberAsync([FromBody]ApplicationUser user, string phoneNumber)
+        {
+            await _userStore.SetPhoneNumberAsync(user, phoneNumber);
+        }
+
+        [HttpPost("setPhoneNumberConfirmed/{confirmed}")]
+        public async Task SetPhoneNumberConfirmedAsync([FromBody]ApplicationUser user, bool confirmed)
+        {
+            await _userStore.SetPhoneNumberConfirmedAsync(user, confirmed);
+        }
+
+        #endregion
+    }
+}
