+++ WebShop/appsettings.json
@@ -1,6 +1,3 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=WebStore;Trusted_Connection=True;"
-  },
   "ClientAddress": "http://localhost:5001"
 }
